Removed the childscene list.
The Childscene list was ultimately flawed. Scenes should never be able
to wantonly be added to other scenes. The idea was sound, allow for
modular scene combinations, but scenes need to be able to return values
results of their actions, and the parent scene needs to know how to
respond, which require real code. Therefore, there's never a place where
a scene can randomly have a bunch of dynamic children, unless a specific
scene is designed to do just that.