Armies are up and running. HQScene is in a moderate state.
Armies exist at all. And can save and load. Additional changes to items
and units to make them savable and loadable. And allowed GridRenderables
to be made without a defined sprite. I may go back on that later, since
that was only done when  was trying to use GSON.

Also also, did work on HQScene.